+++ devenv-module.nix
@@ -148,7 +148,7 @@ that is defined in flake-module.nix
                 ihpPackageNames = [
                     "ihp-ide" "ihp-hsx" "ihp-schema-compiler"
                     "ihp-postgres-parser" "ihp-context" "ihp-pagehead"
-                    "ihp-log" "ihp-modal" "ihp-mail"
+                    "ihp-modal" "ihp-mail"
                     "ihp-migrate" "ihp-openai" "ihp-ssc" "ihp-graphql"
                     "ihp-datasync-typescript" "ihp-sitemap"
                     "ihp-job-dashboard" "ihp-imagemagick"
@@ -178,7 +178,7 @@ that is defined in flake-module.nix
                 ihpPackageNames = [
                     "ihp-ide" "ihp-hsx" "ihp-schema-compiler"
                     "ihp-postgres-parser" "ihp-context" "ihp-pagehead"
-                    "ihp-log" "ihp-modal" "ihp-mail"
+                    "ihp-modal" "ihp-mail"
                     "ihp-migrate" "ihp-openai" "ihp-ssc" "ihp-graphql"
                     "ihp-datasync-typescript" "ihp-sitemap"
                     "ihp-job-dashboard" "ihp-imagemagick"
@@ -493,7 +493,6 @@ that is defined in flake-module.nix
                     docPackages = with pkgs.ghc; [
                         ihp-with-docs
                         ihp-mail
-                        ihp-log
                         ihp-modal
                         ihp-ssc
                         # ihp-hsx ships with `doHaddock = false` in its default.nix