+++ ihp-pglistener/default.nix
@@ -1,5 +1,5 @@
 { mkDerivation, aeson, async, base, bytestring, containers
-, hashable, hasql, hasql-notifications, hspec, ihp-log, lib
+, hashable, hasql, hasql-notifications, hspec, fast-logger, lib
 , safe-exceptions, string-conversions, text, unagi-chan
 , unordered-containers, uuid
 }:
@@ -9,12 +9,12 @@ mkDerivation {
   src = ./.;
   libraryHaskellDepends = [
     aeson async base bytestring containers hashable hasql
-    hasql-notifications ihp-log safe-exceptions string-conversions text
+    hasql-notifications fast-logger safe-exceptions string-conversions text
     unagi-chan unordered-containers uuid
   ];
   testHaskellDepends = [
     aeson async base bytestring containers hashable hasql
-    hasql-notifications hspec ihp-log safe-exceptions
+    hasql-notifications hspec fast-logger safe-exceptions
     string-conversions text unagi-chan unordered-containers uuid
   ];
   homepage = "https://ihp.digitallyinduced.com/";