+++ ihp/IHP/Job/Runner/MainLoop.hs
@@ -15,7 +15,7 @@ import qualified System.Posix.Signals as Signals
 import qualified System.Exit as Exit
 import qualified IHP.PGListener as PGListener
 import Control.Monad.Trans.Resource
-import qualified IHP.Log as Log
+import System.Log.FastLogger (toLogStr)
 import Control.Concurrent.STM (atomically, writeTBQueue)
 
 -- | Used by the RunJobs binary
@@ -29,9 +29,8 @@ dedicatedProcessMainLoop jobWorkers = do
     threadId <- Concurrent.myThreadId
     exitSignalsCount <- newIORef 0
     workerId <- UUID.nextRandom
-    let logger = ?context.logger
 
-    Log.info ("Starting worker " <> tshow workerId)
+    ?context.logger (toLogStr ("Starting worker " <> tshow workerId))
 
     -- The job workers use their own dedicated PG listener as e.g. AutoRefresh or DataSync
     -- could overload the main PGListener connection. In that case we still want jobs to be
@@ -49,7 +48,7 @@ dedicatedProcessMainLoop jobWorkers = do
 
             liftIO waitForExitSignal
 
-            liftIO $ Log.info ("Waiting for jobs to complete. CTRL+C again to force exit" :: Text)
+            liftIO $ ?context.logger (toLogStr ("Waiting for jobs to complete. CTRL+C again to force exit" :: Text))
 
             -- Stop subscriptions and poller already
             -- This will stop all producers for the queue
@@ -69,7 +68,7 @@ dedicatedProcessMainLoop jobWorkers = do
             liftIO $ async do
                 waitForExitSignal
 
-                Log.info ("Canceling all running jobs. CTRL+C again to force exit" :: Text)
+                ?context.logger (toLogStr ("Canceling all running jobs. CTRL+C again to force exit" :: Text))
 
                 forEach processes \JobWorkerProcess { dispatcher = (dispatcherKey, _) } -> do
                     release dispatcherKey  -- cancels dispatcher, whose finally cancels all workers